Chore/refactor

+++ src/common/global_types.rs
@@ -1,7 +1,7 @@
 pub type Pressure = f64;
 pub type Depth = f64;
-pub type Seconds = usize;
-pub type Minutes = usize;
+pub type Seconds = u64;
+pub type Minutes = u64;
 pub type GradientFactor = u8;
 pub type GradientFactors = (u8, u8);
 pub type MbarPressure = u16;
